proxy flag added * main.go simplified * network related lines moved to src/client.go
